Rename group
Change group and package to klaxon.klaxon. Will change again to <domain>.klaxon.klaxon once I get a webpage
