+++ build.gradle
@@ -14,7 +14,7 @@ apply plugin: 'eclipse'
 apply plugin: 'maven-publish'
 
 version = '0.2'
-group = 'klaxon' // http://maven.apache.org/guides/mini/guide-naming-conventions.html
+group = 'klaxon.klaxon' // http://maven.apache.org/guides/mini/guide-naming-conventions.html
 archivesBaseName = 'klaxon'
 
 sourceCompatibility = targetCompatibility = compileJava.sourceCompatibility = compileJava.targetCompatibility = '1.8' // Need this here so eclipse task generates correctly.
