+++ src/main/java/klaxon/klaxon/Klaxon.java
@@ -1,35 +1,27 @@
-package klaxon;
+package klaxon.klaxon;
 
-import klaxon.init.BlockInit;
-import klaxon.init.ItemInit;
-import klaxon.init.SoundInit;
-import net.minecraft.block.Block;
-import net.minecraft.block.Blocks;
+import klaxon.klaxon.init.BlockInit;
+import klaxon.klaxon.init.ItemInit;
+import klaxon.klaxon.init.SoundInit;
 import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.event.RegistryEvent;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
-import net.minecraftforge.fml.InterModComms;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
-import net.minecraftforge.fml.event.lifecycle.InterModEnqueueEvent;
-import net.minecraftforge.fml.event.lifecycle.InterModProcessEvent;
 import net.minecraftforge.fml.event.server.FMLServerStartingEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import java.util.stream.Collectors;
-
 // The value here should match an entry in the META-INF/mods.toml file
-@Mod("klaxon")
+@Mod("klaxon/klaxon")
 public class Klaxon
 {
     // Directly reference a log4j logger.
     public static final Logger LOGGER = LogManager.getLogger();
     //Set the mod id so I don't have to remember it
-    public static String MOD_ID = "klaxon";
+    public static String MOD_ID = "klaxon/klaxon";
 
     //Main class, main method, the entry point to the mod
     public Klaxon() {
